PB-32 feat: Add base class for models and documentable interface
The code changes introduce a new base class `AggregateRoot` for all models to be saved in the database. This class provides a unique identifier (`id`) and getter/setter methods for it. Additionally, a new interface `Documentable` is added, which represents models that can be converted to a `Document` object. These changes aim to improve the organization and consistency of the codebase.