PB-31 feat/change database to mongodb (#17)
* Pb 32 feat/aggregate root for all documentable models (#15)

* PB-32 feat: Add base class for models and documentable interface

The code changes introduce a new base class `AggregateRoot` for all models to be saved in the database. This class provides a unique identifier (`id`) and getter/setter methods for it. Additionally, a new interface `Documentable` is added, which represents models that can be converted to a `Document` object. These changes aim to improve the organization and consistency of the codebase.

* PB-32 refactor: add super class AggregateRoot for objects save in mongodb

* PB-32 refactor: update delete method parameters to use String instead of int

The code changes modify the `delete` method in the `OrderController` class and the `erase` method in the `InMemoryOrderRepository` class to accept a `String` parameter instead of an `int` parameter. This change ensures consistency with the `OrderRepository` interface, which also uses a `String` parameter for deletion. The purpose of this modification is to align the method signatures and improve code readability.

* PB-33 feat: add MongoRepository and MongoOrderRepository classes (#16)

The code changes introduce the `MongoRepository` class, which represents the repository for the mongo database. It provides methods for persisting, searching, and deleting aggregate roots. Additionally, the `MongoOrderRepository` class is added, which extends the `MongoRepository` and implements the `OrderRepository` interface. This class specifically handles the persistence and retrieval of `Order` objects in the MongoDB database. These changes aim to improve the organization and structure of the codebase.