PB-38 feat/delete an order in views (#19)
* fix: update pom.xml dependencies for servlet and servlet-api

The code changes in the `pom.xml` file update the dependencies for the servlet and servlet-api. The `jakarta.servlet.jsp.jstl` dependency is removed, and the `javax.servlet-api` dependency is added with version `4.0.1` and scope `provided`. This update aims to align the project with the latest servlet specifications and improve compatibility with the web application.

* fix: add documentation for MongoClientFactory and public class

* feat: add OrderView servlet for deleting orders

The code changes add a new servlet called `OrderView` that handles the deletion of orders. This servlet receives a request with an order ID parameter and uses the `OrderController` and `OrderRepository` to delete the corresponding order from the MongoDB database. The purpose of this addition is to provide functionality for deleting orders in the web application.

* feat: add StringUtils class for UUID validation

The code changes introduce a new utility class called `StringUtils` that provides a method `isUuid` for checking if a string is a valid UUID. This class aims to improve the codebase by centralizing the UUID validation logic and promoting code reusability.