PB-26 refactor/changes some code for enum (#26)
* PB-34 feat/funcionatility with controllers and database (#25)

* PB-38 feat/delete an order in views (#19)

* fix: update pom.xml dependencies for servlet and servlet-api

The code changes in the `pom.xml` file update the dependencies for the servlet and servlet-api. The `jakarta.servlet.jsp.jstl` dependency is removed, and the `javax.servlet-api` dependency is added with version `4.0.1` and scope `provided`. This update aims to align the project with the latest servlet specifications and improve compatibility with the web application.

* fix: add documentation for MongoClientFactory and public class

* feat: add OrderView servlet for deleting orders

The code changes add a new servlet called `OrderView` that handles the deletion of orders. This servlet receives a request with an order ID parameter and uses the `OrderController` and `OrderRepository` to delete the corresponding order from the MongoDB database. The purpose of this addition is to provide functionality for deleting orders in the web application.

* feat: add StringUtils class for UUID validation

The code changes introduce a new utility class called `StringUtils` that provides a method `isUuid` for checking if a string is a valid UUID. This class aims to improve the codebase by centralizing the UUID validation logic and promoting code reusability.

* Pb 36 create view and controller to add orders (#20)

* PB-36 feat:add order

* PB-36 feat: create view controller add orders

---------

Co-authored-by: Jean Valencia <72704457+Jean-Dv@users.noreply.github.com>

* PB-34 feat: add display all orders in the table with database (#21)

* PB-35 create view and controller to display orders

* PB-35 feat: connect view controller to display orders

* fix: delete unnecessary import

* fix: modify endpoint name to orders

* refactor: update hrefSection and hrefTitle values in orders.jsp

* feat: Add delete button to orders.jsp

The code changes in `orders.jsp` add a delete button to each order row. This button allows users to delete an order directly from the web page.

* style: change a spacing in one line of jsp

---------

Co-authored-by: Jean Valencia <72704457+Jean-Dv@users.noreply.github.com>
Co-authored-by: Jean-Dv <jean.valencia01@proton.me>

* PB-38 feat/delete an order in list of jsp (#22)

* refactor: move the scripts of javascript for orders.jsp

* fix: improve UUID validation in StringUtils

* feat: add delete functionality in orders view

* refactor: remove unnecessary code in OrderView

* interface is added to edit orders and send information to the controller (#24)

Co-authored-by: Jean Valencia <72704457+Jean-Dv@users.noreply.github.com>

---------

Co-authored-by: Bibian Corredor <138538175+BibianCo@users.noreply.github.com>
Co-authored-by: Diego Combariza <136619052+DiegoCombariza11@users.noreply.github.com>
Co-authored-by: juan-Riapira <138538939+juan-Riapira@users.noreply.github.com>

* fix: change method to call from document to model on java

* feat: add default status warehouse exit when add an order

* refactor: update select element styles in editorder.jsp

* refactor: update order status display in orders.jsp and cash on delivery

---------

Co-authored-by: Bibian Corredor <138538175+BibianCo@users.noreply.github.com>
Co-authored-by: Diego Combariza <136619052+DiegoCombariza11@users.noreply.github.com>
Co-authored-by: juan-Riapira <138538939+juan-Riapira@users.noreply.github.com>