PB-47 feat/graphic interfaz for settings (#37)
* chore: delete one file

* fix: update methods for compile the project

* feat: Add SettingsView servlet for displaying settings

* feat: Add jstl dependency to pom.xml

The code changes in the `pom.xml` file add the `javax.servlet:jstl:1.2` dependency. This dependency is required for using JSTL (JavaServer Pages Standard Tag Library) in the project.

* feat: method to send data from the form in the view

* feat: Add settings.jsp page for displaying settings

The code changes add a new JSP page called `settings.jsp` that is responsible for displaying the settings of the application. This page includes HTML markup and JSTL tags to render the settings form and show any success messages. This addition enhances the user interface by providing a dedicated page for managing application settings.

* fix: modified sidemenu.jsp for the actual menu completed

* feat: update OrderView to use source address from settings

The code changes in `OrderView.java` modify the logic for creating an `Order` object. Instead of using an empty string for the source address, the code now retrieves the source address from the settings using the `SettingsController`. This enhancement ensures that the `Order` object is created with the correct source address, improving the accuracy of the order information.