PB-60 feat/show map with routes (#61)
* PB-60 fix: update geocoding address to use correct spelling for Boyacá

* PB-60 feat: add Leaflet map to display routes in editstate.jsp

The code changes in `editstate.jsp` add the necessary HTML markup and JavaScript code to display a Leaflet map on the page. This map is used to show the routes of the order. The Leaflet library is loaded from a CDN and the map is initialized with a default view and zoom level. Functions are added to add markers and polygons to the map based on the data received from the server. The map is displayed below the order details section.

Co-authored-by: Juan-Fernandez <juan.fernandez07@uptc.edu.co>
Co-authored-by: juan-Riapira <138538939+juan-Riapira@users.noreply.github.com>

* PB-60 refactor: add PropertiesUtils class for loading properties

* PB-60 refactor: change hardcoding to app properties correctly

* PB-60 fix: update loadProperties to local.properties ignore

---------

Co-authored-by: Juan-Fernandez <juan.fernandez07@uptc.edu.co>
Co-authored-by: juan-Riapira <138538939+juan-Riapira@users.noreply.github.com>