PB-13 fix: update MongoDB database property to use default value if not specified
The code changes in AlgorithmStar.java, MongoRepository.java, DigraphLoader.java, MongoOrderRepository.java, and Main.java update the database property to use a default value of "logistics" if it is not specified in the application.properties file. This ensures that the application can connect to the MongoDB database even if the property is missing or empty.