PB-13 feat/show the most optimal route (#62)
* Pb 59 feat/get optimal routes and send them to the server (#59)

* PB-59 feat:Obtain the most optimal route and its alternative routes from the database, send them to the server

* fix: OrderView class imports that were not being used

* Pb 59 feat/get optimal routes and send them to the server (#60)

* PB-59 feat:Obtain the most optimal route and its alternative routes from the database, send them to the server

* fix: OrderView class imports that were not being used

* PB-59 fix:add a new RouterView class where the optimal routes are shown and eliminate lines in the OrderView

* PB-60 feat/show map with routes (#61)

* PB-60 fix: update geocoding address to use correct spelling for Boyacá

* PB-60 feat: add Leaflet map to display routes in editstate.jsp

The code changes in `editstate.jsp` add the necessary HTML markup and JavaScript code to display a Leaflet map on the page. This map is used to show the routes of the order. The Leaflet library is loaded from a CDN and the map is initialized with a default view and zoom level. Functions are added to add markers and polygons to the map based on the data received from the server. The map is displayed below the order details section.

Co-authored-by: Juan-Fernandez <juan.fernandez07@uptc.edu.co>
Co-authored-by: juan-Riapira <138538939+juan-Riapira@users.noreply.github.com>

* PB-60 refactor: add PropertiesUtils class for loading properties

* PB-60 refactor: change hardcoding to app properties correctly

* PB-60 fix: update loadProperties to local.properties ignore

---------

Co-authored-by: Juan-Fernandez <juan.fernandez07@uptc.edu.co>
Co-authored-by: juan-Riapira <138538939+juan-Riapira@users.noreply.github.com>

* PB-13 fix: update MongoDB database property to use default value if not specified

The code changes in AlgorithmStar.java, MongoRepository.java, DigraphLoader.java, MongoOrderRepository.java, and Main.java update the database property to use a default value of "logistics" if it is not specified in the application.properties file. This ensures that the application can connect to the MongoDB database even if the property is missing or empty.

---------

Co-authored-by: Bibian Corredor <138538175+BibianCo@users.noreply.github.com>
Co-authored-by: Juan-Fernandez <juan.fernandez07@uptc.edu.co>
Co-authored-by: juan-Riapira <138538939+juan-Riapira@users.noreply.github.com>