PB-64 fix: add error message for Google Maps in editorder.jsp
The code changes in editorder.jsp add an error message for Google Maps. If there is an error with the Google Maps API, the error message will be displayed in red on the page. This improves the user experience by providing feedback when there is an issue with the Google Maps functionality.